Hide destroy button for regular users

+++ app/helpers/application_helper.rb
@@ -3,9 +3,13 @@ module ApplicationHelper
     return nil if current_user.nil?
 
     edit = link_to('Edit', url_for([:edit, item]), class: 'btn btn-primary')
-    del = link_to('Destroy', item, method: :delete,
-                                   data: { confirm: 'Are you sure?' },
-                                   class: 'btn btn-danger')
+
+    if admin_user
+      del = link_to('Destroy', item, method: :delete,
+                                     data: { confirm: 'Are you sure?' },
+                                     class: 'btn btn-danger')
+    end
+
     raw("#{edit} #{del}")
   end
 end
